+++ src/com/github/mjkuranda/retrographicsmanager/Main.java
@@ -1,0 +1,12 @@
+package com.github.mjkuranda.retrographicsmanager;
+
+/***
+ * @author mjkuranda
+ */
+public class Main {
+
+    public static void main(String[] args) {
+        RetroGraphicsManager app = RetroGraphicsManager.getInstance();
+        app.start();
+    }
+}