+++ src/com/github/mjkuranda/retrographicsmanager/RetroGraphicsManager.java
@@ -9,10 +9,35 @@ public class RetroGraphicsManager {
     public static final String VERSION = "3.1.0";
     public static final String RELEASE = "2023-05-17";
 
+    private static RetroGraphicsManager app;
     private final ApplicationManager manager;
 
-    public RetroGraphicsManager() {
-        manager = ApplicationManager.getInstance();
+    public static RetroGraphicsManager getInstance() {
+        if (app == null) {
+            app = new RetroGraphicsManager();
+        }
+
+        return app;
+    }
+
+    private RetroGraphicsManager() {
+        manager = new ApplicationManager();
+    }
+
+    private class ApplicationManager {
+        private boolean isRunning;
+
+        ApplicationManager() {
+            this.isRunning = false;
+        }
+
+        void setRunning(boolean isRunning) {
+            this.isRunning = isRunning;
+        }
+
+        boolean isRunning() {
+            return isRunning;
+        }
     }
 
     public void start() {
@@ -20,7 +45,7 @@ public class RetroGraphicsManager {
         run();
     }
 
-    private void stop() {
+    public void stop() {
         manager.setRunning(false);
     }
 
@@ -36,28 +61,10 @@ public class RetroGraphicsManager {
         while (manager.isRunning()) {
             System.out.print("> ");
             lineArgs = scan.nextLine().split(" ");
-
-            if (ifTerminate(lineArgs)) {
-                stop();
-
-                continue;
-            }
-
             executeCommand(lineArgs);
         }
 
         System.out.println("Exit");
-    }
-
-    private boolean ifTerminate(String[] lineArgs) {
-        if (lineArgs.length == 0) {
-            return false;
-        }
-
-        return switch(lineArgs[0]) {
-            case "exit", "end", "term", "terminate" -> true;
-            default -> false;
-        };
     }
 
     private void executeCommand(String lineArgs[]) {