+++ src/com/github/mjkuranda/retrographicsmanager/commandsystem/CommandFactory.java
@@ -5,11 +5,10 @@ import com.github.mjkuranda.retrographicsmanager.commandsystem.commands.*;
 public class CommandFactory {
 
     public static Command get(String[] lineArgs) {
-        Command c = switch (lineArgs[0]) {
+        return switch (lineArgs[0]) {
             case "list", "ls" -> new ListCommand(lineArgs);
+            case "exit", "end", "terminate", "term" -> new TerminateCommand();
             default -> new UnknownCommand(lineArgs);
         };
-
-        return c;
     }
 }